RefetchData - Move component to `src`. - Export component from `index`. - Update story to reflect import from `index`.
