+++ src/index.js
@@ -21,6 +21,7 @@ export { default as List } from './list'
 export { default as Loader } from './loader'
 export { default as QuickFilters } from './quick-filters'
 export { default as Radio } from './radio'
+export { default as RefetchData } from './refetch-data'
 export { default as StyledCardContainer } from './styled-card-container'
 export { default as Switch } from './switch'
 export { default as Table } from './table'
